BUG #15093: Fixed issue with stereo rendering.
The problem was that the StereoType set on vtkRenderWindow was
incorrect. It ended up picking Crystal Eyes incorrectly. The problem was
due to the the "StereoType" property being reset to default by
vtkSMParaViewPipelineController. This fixes that.

Also added tests to test setting of stereo modes to catch this in
future.

Change-Id: I20d7227d2059be007b8ff2d1a6abffa349679345