BUG #15060. Fixed issue when using TemporalShiftScale.
The problem was that when the current timesteps changed, we didn't snap
the current time to the new timesteps. That meant that we continued to
show the same old timestep even though the pipeline timesteps changed.
We fix this behavior by ensuring that we snap to a valid timestep when
the timesteps change and the animation is in "snap to timesteps" mode.

Change-Id: Ib6d1c1ca22eedf84cc8e9aa36592ad403d10a172