HACK: Fix duplicate entries in Color By.
We sometimes end up with duplicate entries. Overcome it for now.
We need a proper fix in vtkSMArrayListDomain/vtkSMRepresentedArrayListDomain/
vtkPVDataInformation.  for now, just do this. To reproduce the problem,
skip the uniquifier check, and the load can.ex2, uncheck all blocks,
and check all sets. The color-by will have duplicated ObjectId,
SourceElementId, and SourceElementSide entries.

Change-Id: I026218602671c6777483c29d1ada6918461c4bd5