Added missing Q_Object macro for a couple qt event players.
The CTK QtTesting repo has some changes that uses the qt metaObject
class name to check whether an event player is already in the players
list. So if an event player class is not declared as Q_Object, the
metaobject will just use its superClass name, and this is why these
paraview event players are missing from the players pool, which
causes many paraview tests to fail.

Change-Id: I75e1a25a55587b68efb59b1c2bfb420fee8c0125