Refactoring slice view/quad view.
This is a refactor of the views that the ParaView UI called "Slice View"
and "Quad View".

This change has several cleanups that should make the views easier to
maintain. The improvements include:

* vtkPVOrthographicSliceView replaces vtkPVQuadView.

Added vtkPVOrthographicSliceView to replace vtkPVQuadView. Quad view
becomes a more understandable "Orthographic Slice View".

* Changed the way basis vectors are specified.

The new implementation simply expects a change-of-basis-matrix and then
the bounds for the dataset in original basis. Both the Slice View and
Orthographic Slice view supports this new basis vector definition
mechanism. vtkPVChangeOfBasisHelper is added to make it easier for
applications to add/retrieve this meta-data.

* vtkShearedWaveletSource replaces vtkShearedCubeSource.

vtkShearedCubeSource is replacted by vtkShearedWaveletSource which uses
the new way of specifying basis vectors.

* vtkPVOrthographicSliceView replaces QuadView.

vtkPVOrthographicSliceView users a single window with multiple renders
to show the orthographic views. This makes it easier for applications to
handle this view as simply a specialization of RenderView. No special
layouting code needed to lay out the internal sub-views.