Separated out EventPlayer from pqViewStreamingBehavior
The pqViewStreamingBehavior contains and instantiates its own event
player to play back its recorded event. The new QtTesting flow does
not support this. The players have to be registered first before
recorded event can be dispatched to the player. So the internal
event class is moved out to be a new class and registerd on its own.

Change-Id: I8c4b48a492a0088a6dcdf8de82313bee8185ec54