Added new vtkAnnotateAttributeDataFilter.
Added new vtkAnnotateAttributeDataFilter that can be used to generate
annotation text using data from any element in the dataset. By picking
the ElementId as a property, we can setup animations where the ElementId
changes over the course of the animation to show different elements from
a dataarray as annotation text!

Change-Id: I3d05f7fdb54e8bfaa9581a0b3f79bcde306686ec