Fixed animating slice-origin in comparative view.
Comparaview view could not animate properties on helper proxies e.g.
Slcie Origin on the Plane proxy for a Slice filter. This was because
pqComparativeVisPanel wasn't using the proper proxy when adding a cue to
animation a helper-proxy property. Fixed that.

Also consolidate code to locate proxy with a given helper proxy in
pqProxy. Both pqAnimationViewWidget and pqComparativeVisPanel now use
that same method.

Change-Id: Iacaf45909919e0d96459bd46badfdfe66ad61b5b