Moved map lookup outside of inner loop
This commit pulls a change that was made on the new multi-block reader
back to be the original reader from which the segment of code was copied.
The optimization is to move the map lookups getting the type and buffer
for the array outside of the inner loop that reads in the points.

Change-Id: I19973e962399db7753286accb335768926837c6c