BUG: Adjust Camera / Elevation button printed warning.
Adjust Camera / Elevation button used to adjust only the position
of the camera until (Position,Focal Point) axis was parallel with
the ViewUp vector when a warning was printed:
vtkOpenGLRenderer (0xcbf02b0): Resetting view-up since view plane normal
is parallel.
Now we modify both Position and ViewUp vector so we avoid this problem.

Change-Id: If07e940e5fd3b3b2d96746850f0e375a80062666