Remove the requirement on using std::type_traits.
Previously we used type_traits, and presumed it was always in the tr1 namespace.
This is really tricky to get right, as depending on the compiler and the flags
given to it the location of type_traits is different. Since the functionality
that we require is fairly simple we just implement it our selves.

Change-Id: I45f0128c6b5b1aa8ed309790da2fc85a63ab1834