Add a cinema export option to coprocessing scripts
The coprocessing plugin gets a new option to export
cinema "spec A" files. For every view saved, the decides
if it will be exported to a cinema store, and if so whether
it should have a static, or spherical camera. Likewise every
slice plane and isocontour gets a track that the user can
specify what values it should range over in the store.

Besides the UI changes, this change brings in an extract
of the "spec A" cinema store python module. Catalyst calls
into it at run time to create the store(s).

Change-Id: Ia8d2eb6af8d286cab845fba9383689c9502d9e02