Added ability to save default color maps for arrays
The existing functionality saved one color map to be applied to ALL
arrays when the color map is first created. This change makes it
possible to save a default color map for an array with a particular
name so that any time an object is first colored by that array, it
will be set to the default for that array.

To enable this feature, added API to vtkSMSettings to save a proxy's
settings under a specified prefix. In the case of color map defaults,
we set the prefix to .lookup_tables.arrays.<array name>. The defaults
for a color map can be queried with the same prefix. Loading the
default color map for an array is done in vtkSMTransferFunctionManager
while saving is taking care of in pqColorMapEditor.

In the JSON settings file, a color map entry for arrays named
"ObjectId" will look like:

{
  "array_lookup_tables" : {
    "ObjectId" : {
      "PVLookupTable" : {
        "Annotations" : [ "1", "1", "2", "2" ],
        "IndexedColors" : [
          0.988, 0.552, 0.349,
          1, 1, 0.749,
          0.6, 0.835, 0.580
        ],
        "RGBPoints" : [
          1, 0.231373, 0.298039, 0.752941,
          1.5, 0.865003, 0.865003, 0.865003,
          2, 0.705882, 0.0156863, 0.14902
        ]
      }
    }
  }
}

Note the the sanitized array name is stored in the JSON file rather
than the original array name. The sanitized name is how color maps are
referenced by vtkSMTransferFunctionManager.

In the GUI, added a button in the color map editor to save the default
for the current array name. Added a new icon for this.

Change-Id: I7bdcbe05b4a7171203aa1fa68f0937e96048a81a