BUG #15247: Restore mouse-wheel.
The wheel was broken in e34444b0515c. The change was cleaning up
vtkPVInteractorStyle. vtkPVInteractorStyle has traditionally been
subclass of vtkInteractorStyleTrackballCamera. Restoring that behaviour.
Ideally, we want to add better support for mouse-wheel. Leaving the
incidental support for now. Also manually disabled wheel in
vtkPVOrthographicSliceView since the wheel conflicts with the user
interactions for moving slices in that view.

Change-Id: I4ae5db174e60f208c64a53ecf2dcf00459d2be7c