Fixed BUG #14953. Properly handle Esc key in Search box.
To fix BUG #14953, we properly handle Esc key in the Search box.
pqSearchBox now eats the Esc key if the Esc key will clear the text in
the Search box and passes it only if the text box is already empty.
Passing it through will cause the parent widget to do its default thing,
e.g. closing of dialogs. As a result the Settings dialog now behaves as
follows: when there is some text in the Search box, hitting Esc will clear it.
If you hit Esc again (or when the Search box) is empty, the dialog does
it's default thing, i.e it closes.

Change-Id: I4b3c47f8056cf5f0311a9e96bda6a0c91173bb5d