Refactoring ParaView interaction code.
ParaView was subclassing vtkRenderWindowInteractor to create
vtkPVGenericRenderWindowInteractor to handle interaction. That piece of code
was potentially derrived from an older implementation of
vtkRenderWindowInteractor and hence did what it did. Current implementation
of vtkRenderWindowInteractor lets the vtkInteractionStyle (and subclasses)
do all the heavy lifting. ParaView did that to some extent (since it has a
vtkPVInteractorStyle), but will was relying on
vtkPVGenericRenderWindowInteractor, vtkPVRenderViewProxy to propage
interaction/still renders and other things. This has been refactored.  ParaView no
longer uses a special vtkRenderWindowInteractor. All logic is handled by observers
on the standard vtkRenderWindowInteractor.

This change was done to make it easier to enable interactivity in pvpython.
Added new API to simple.py to start interact with a view in pvpython.

Change-Id: I9725be948174b12c7f306c1d2fdbcfaac2d352ee