vtkMPI: include vtkMPI before making libraries
Due to CMake's scoping of variables and functions, vtkMPI needs to be
included at each location vtk_mpi_link is used. However, there are
important things vtkMPI brings in such as include directories and
preprocessor definitions, so for them to take effect, they must be
available when vtk_module_library is called.

Change-Id: I1531b55d1adcb9e01441dfe9fbd12905193ba9a0