Simplify Python interaction API.
Previously, if Render() was called before Interact(), we couldn't
support that since the interactor cannot be created safely after the
window has been created. To address that issue, we now update simple.py
to create the interactor right after the view is created and before the
render is called. This will address most use cases.

Change-Id: Ic39f22959756ad6a7ef4e875ef11be25b15f6b0e