Widgets didn't get the interactor when loading state files.
Now that the interactor is setup by the application after the view has
been created, when loading state files that had the scalar bar widget,
the scalar bar widget ended up not seeing the interactor set on the view
since the scalar bar was added to the view before the view's interactor
was setup. This fixes that issue by making vtkPVRenderView fire
ModifiedEvent everytime the interactor is changed/setup and making
vtk3DWidgetRepresentation monitor that event to update the interactor on
the interal vtkAbstractWidget.

Change-Id: I994cde4153b4ca60c4e1e2fc7b2eb7cdbeb423a2