Fix issues with CinemaTest with interactors.
Creating a default interactor was causing incorrect cleanup for Cinema
test. The issue was arising due to incorrect OpenGL clean in display
list painter. Since that code is going to be deprecated, overcoming the
issue by making vtkSMViewProxy create the
vtkGenericRenderWindowInteractor in batch mode and updating CinemaTest
to use pvbatch instead of pvpython.

Change-Id: Ie6edd222d5fe1b4ae328affbe585747660a30b24