BUG #15298. Fixes exporters in Python.
Proxies that were not made class-types for in updateModule/_createModule calls
ended up not have the Python property API essential for SetProperties(...) call
to work. One way is to ensure _createModule/updateModule initializes all
proxy groups -- that's bound to fail if new groups get added. To make it
more robust, instead we not create a class-type on demand (if one
doesn't already exist) in _getPyProxy().

This fixes the issue reported in BUG #15298.

Change-Id: Ib7099f0395c74c699878855378c7555cd5ca118a