Make pqServerResource track the pqServerConfiguration.
pqServerResource now tracks the pqServerConfiguration that it was born
from. This makes it possible for the UI to know which configuration was
used to connect to the specified server.

If the pqServerResource was not created from a pqServerConfiguration
e.g. when ParaView is started with -url command line argument. This
change still works, since it just creates an empty/default
pqServerConfiguration for such a pqServerResource.

Change-Id: I14709287fedbccb0d5b041775b75a91a96461bc6