BUG #15318. Added ability to pick different client/server stereo types.
Now, when ParaView is run with stereo enabled  i.e. "paraview --stereo",
the Properites panel allows the user to pick StereoType for the client
and server processes separately. There properties are only shown in the
panel when stereo is enabled (using the command line on the client
process). By default, the server stereo type is set to track the client
stereo type.

Change-Id: Ie6eabe024a7287e0d856870aa3ad93d806adc6e6