Added GeodesicMeasurement plugin
The plugin is meant to be compiled and shipped with ParaView, but not
necessarily enabled by default. It exposes a filter that is used to
compute piece-wise geodesic paths between a set of input points.  The
first input to this filter is a polygonal surface on which geodesic
measurements should be taken. Points in the second input are used to
define the endpoints. Geodesic paths are computed between consecutive
points in the endpoints input and form a partial loop. Options for
completing the loop with another geodesic or with a straight line are
available.

Change-Id: Iee70117176567aedba3f2e57057d0dcca4fb5485