Integrate Cray ATP
Abnormal Termination Processing (ATP) is a system that monitors Cray
system user applications. If an application takes a system trap, ATP
performs analysis on the dying application. All stack backtraces of
the application processes are gathered into a merged stack backtrace
tree and written to disk as the file, atpMergedBT.dot. The stack
backtrace tree for the first process to die is sent to stderr as is
the number of the signal that caused the application to fail.  If
Linux core dumping is enabled (see ulimit or limit in your shell
documentation), a heuristically selected set of processes also dump
their cores. To use ATP, the ATP must be linked into the application
and the shell environemtn variable ATP_ENABLED=1 must be set prior to
launching the program with aprun. This patch adds a CMake module to
locate APT libraries and set up link flags. The new module will be
used to link ATP lib into the command line executables if the CMake
variable PARAVIEW_USE_ATP is ON and ATP is found.

Change-Id: I29e817338dd0f199b55d395019d58a6f23f02be4