BUG #15352. Fix python initialization in pvbatch.
pvbatch was not initializing the static python modules on ranks other
than 0. Which meant that the programmable filter could not execute on
those other ranks since it failed to file the static Python modules.
This fixes that issue. Python is now initialized on all ranks.

This manifested as a failure in test
pvpythonPython-MPI-Batch-TestAnnotateAttributeData in static builds with
Python/MPI enabled.

Change-Id: I55bfdbfc31056e385f310c37f26343976b04d30e