Adding support for reading image stacks.
Added a new subclass to vtkFileSeriesReader named
vtkImageFileSeriesReader which can be used for vtkImageReader2
subclasses. It has a ivar, ReadAsImageStack, which when set to true
will cause the reade to read the files as a stack of images rather than
a temporal file series.

Updating JPEG, TIFF and PNG readers to read in the image-stack mode, by
default.

Developer note:
Ideally, this change should be integrated in vtkFileSeriesReader itself.
However, the vtkFileSeriesReader class hierarchy is a little confusing
and would need some refactoring to clean it up. I am taking the path of
least resistance for now.