Update submodule URLs to use Gitlab.
Updating repository URLs for submodules that have migrated to
gitlab.kitware.com to use those URLs instead of the old ones.

After this change, users will need to call `git submodule sync` to
ensure that the new URLs are used for existing submodules. This is not
needed if cloning a fresh copy.