Fix 'default_values' specified in ExposedProperties.
If default_values are specified in ExposedProperties XML elements to
override the default on the subproxy's properties, they weren't working.
Since the new default values would not align with the VTK side. The
issue was that ExposedProperties elements are not parsed on SI side and
hence the server-side doesn't know about the updated defaults. Fixed
that by marking update properties as modified, explicitly. That way,
the new values will get pushed in the next UpdateVTKObjects() call.