Added new annotation for axes grid.
Added new set of classes for rendering the axes grid. vtkGridAxes3DActor
is the new generation of the Cube-Axes annotation. vtkGridAxes3DActor is
designed to render a grid for a scene. It uses vtkAxis for label
placement strategy and uses vtkTextActor for rendering all the text
labels and titles. Despite using vtkTextActor (which is a 2D actor), the
vtkGridAxes3DActor is a vtkProp3D i.e. it renders in a 3D scene (which
certain caveats).