Add vtkGridAxes3DActor to vtkPVRenderView.
ParaView's Render View now uses vtkGridAxes3DActor to show the axes
grid. To support vtkGridAxes3DActor in parallel without having to
unstage IceT's compositing, we now have 3 renderers in vtkPVRenderView,
a new one for background. The grid axes actor renders background text in
that renderer so that it is drawn behind composited results when the
client gets the image delivered to it.

Caveat: we don't support vtkGridAxes3DActor in tile-display mode and
cave mode since the background renderer doesn't work with how we render
in these modes. Leaving that out for now.