Updates to change-of-basis support for grid axes.
The grid axes now works with change-of-basis matrix. We now use a
behavior to setup axis titles, and such other MantId specific
customizations. While this is not through, this lays the ground work to
start removing explicit code in the rendering pipeline to handle such
meta-data and instead rely on the pqModelTransformSupportBehavior to
setup it correctly.