Adding 'command' to <SubProxy> element.
It's often that we need to pass the VTK object for a subproxy to the
container proxy using a Set* method. We ended up subclassing vtkSIProxy
for such cases. Adding support to simply specify the command to use to
on the XML definition itself. Hence all such subclasses become
unnecessary!