Handle timesteps from programmable source/filter.
If programmable source/filter is setup to produce timesteps, the
ParaView UI never realizes that. This happens because the source/filter
proxy didn't have the TimeStepValues property. Added that. Now, if the
filter/source produces timesteps, ParaView UI will know about it. If it
doesn't, things should continue to work as before.