BUG #15407 Fix loading states with different timesteps.
This is not-so-elegant way to address BUG #0015407. What's happening is
that when state is being loaded the timekeeper gets updated timesteps which
are then correctly propagated to the animation scene, which then,
correctly (I might add) updates the start and end times. However all this
happends before the "EndTime"/"StartTime" properties from the XML state are
loaded.  Hence they get set and loose the values just updated by the
application, instead use the old values from the state files.

This change will will prune the EndTime/StartTime XML elements from the state
being loaded if they should not be loaded.