qt5: ensure Qt5 targets are in-scope
When found in the CMakeLists.txt, they are not available when
transitively linked (namely in plugins or the applications). When found
in the module.cmake, they are scoped to the VTK subdirectory and not
available back in ParaView. That scope also loses the ${module}_ENABLED
variables, so we can't even find the components based on those
variables. Instead, rely on the knowledge that if Qt is enabled, these
modules will be enabled via the application.