Fixed bug when coloring multiblock data set by field data
Added test for coloring multiblock data set by field data. The
multiblock data set consists of vtkImageData leaves. It was failing
when running in client/server mode with the server run via mpirun with
two or more processors.

The problem was located in
vtkMultiProcessControllerHelper::MergePieces.  Enabling copying of
field data in the vtkAppendCompositeDataLeaves filter used to merge
leaves in multiblock data sets fixes the problem exposed by the test.