ExternalData: add a README file
This file is used as a sentinel file for enabling testing or not. The
git repository will always have it, but now that ParaView will be
shipping a separate data tarball, this file will be in that one. The
CMake code determines the default for testing based on its existence.
This will allow everything to work from the tarballs without hitting the
network by default.