BUG: Settings crash when vector property number of elements changes
When a repeatable vector property has one element and it is saved via
vtkSMSettings, there was a crash if the vector property number of
elements changed to more than one element and then was saved to
vtkSMSettings again. This was caused because we store single-element
properties as raw values in the JSON where as multi-element properties
are saved as JSON arrays. In the JsonCpp library, attempting to resize
a raw value node throws an exception. Added a fix for this problem and
modified the vtkSMSettings test to exercise it.