Added vtkPVRepresentedArrayListSettings class
This class is a singleton that contains settings for the
RepresentedArrayListDomain that is used to decide which arrays are
available for coloring. There is currently one setting that is a
variable-size string vector of regular expressions. Any array name
that matches one of these regular expressions is not added to the
color array list when the list is built.

Added support for filtering by regular expressions in
vtkSMRepresentedArrayListDomain.