Don't render client-side 3D geometry when remote rendering.
Parent commit adds a test that fails to demonstrate the problem
(TestCompositedGeometryCulling). The issue was that the composited
renderer always renders the client-side geometries even when remote
rendering is enabled. This change adds a culler to cull the props which
will be rendered remotely.

Also added logic to vtkPVRenderView to robustly say if remote rendering
is being employed.