BUG #15437: Fix multipiece selection.
This reverts part of 54493a8f. The change to make vtkCompositeIndex a
field-data array is what broke the selection. We cannot make
vtkCompositeIndex field data since a multi-piece dataset is squashed
into a single dataset before passing the geometry on to the mapper.

Also updated baselines for SurfaceLIC tests since they were using the
vtkCompositeIndex for coloring.