BUG #15503: Fix collection information about field data arrays.
When collecting information about field data arrays in parallel, we
accumulated the number of tuples by simply adding them. This resulted in
vtkSMArrayListDomain::FieldArrayHasOneTuplePerCompositeDataSetLeaf()
check to fail. This check limits which field-data arrays are shown in
the color-by drop-down box. This commit fixes the number of tuples
accumulation logic to use the max value for FieldData arrays instead of
sum. This is reasonable for the way field data is reduced in parallel
when rendering on the client process.