BUG #15306: Add support for "radius" when click-selecting a point.
Added a new advanced option on the "Settings | Render View" tab, named
"Point Picking Radius" that can be used to specify a radius (in pixels)
to use when click-selecting a point in the render view using "Select
Points On". This radius is only used when you "click-to-select" or when
using "interactive selection". Hence, if you click-n-drag to produce a
sleecttion region, this radius won't be regarded since it's assumed that
the user is indeed interested in selecting points only in that region.