Added rescale data on visibility change mode
Changed label from "Automatically rescale color map range to fit data"
to "Lock Data Range". The value of this renamed label correponds
exactly to the underlying property value, whereas with the old name it
was the exact opposite value.

Added concept of range for categorical color mode. The Annotations are
still defined in the UI as before and can be considered a database of
annotated values and colors, while only annotated values in the range
of the dataset(s) are displayed in the color legend. The range can be
set by either

  1) selecting a pipeline object and clicking the "Rescale to Data
  Range" button, or

  2) enabling the "Rescale On Visibility Change" option in the color
  map and changing the visibility of a pipeline object.

Since the "range" concept now applies in both the categorical color
mode and scalar mapping mode, the "Lock Data Range" option is always
exposed.

Created a vtkPVDiscretizableColorTransferFunction class that is a
custom color transfer function for ParaView. It exposes new methods to
set all annotations, annotated values, and indexed colors, and passes
a subset of "active" annotations and indexed colors on to the
annotation API in vtkDiscretizableColorTransferFunction. This is the
mechanism used to implement the "range" concept to categorical colors
displayed in the color legend.

Added tests for resetting transfer functions on visibility change and
updates of annotated values in color legend.