BUG #15512: Don't show components for arrays by default in QuartileView.
The problem stems from the fact that "Plot Selection Over Time" filter
likes to split multi-component arrays into multiple arrays with single
component. As a result, the logic in vtkSMChartSeriesSelectionDomain to
not show array components by default, but only plot their magnitude
fails. This commit adds extra logic to try to determine if the array
name for a single component array could still be extracted from a multi
component array. If so, we don't show it by default. This isn't robust,
nor is it elegant, but it gets the job done.