Add editable configurable lights to manta view.
We still disable and ignore the lightkit and default to a single
light headlight. Now the user can add new lights control
their color and intensity, and make them be either them be either
headlights, camera relative lights, or worldspace lights.
Note: I'ld like to use the widget's enable checkbox but it
depends on apply so it doesn't work smoothly.