Fix filtering of field data from vtkSMArrayListDomain.
When adding ability to limit arrays to color with in commit 8739bf2a8,
we inadvertently started filtering field data arrays for all array list
domain including the properties on filters like AnnotateGlobalData.
These filters indeed work with field data arrays have a tuple count
independent of the block count. This commit fixes that change to only
add the tuple count filtering logic to vtkSMRepresentedArrayListDomain
used for properties to pick arrays to color with.