BUG #15322: Fixes global data annotations when using mode shapes.
When using mode-shapes, the timesteps no longer match the number of
tuples in the field data arrays. Hence, the logic to extract the
current timestep equivalent tuple from a global variable fails. We
should really use the 'mode_shape' to index in such cases. To support this, we
needed to add a new variable to vtkExodusIIReader named
'mode_shape_range' which provides the info about the mode shape range to
the vtkAnnotateGlobalData filter. Then, if the number of tuples matches
this range, we use the 'mode_shape' to index into the array. This fixes
the reported issue.