BUG #15320: Fix time precision descrepancies.
The reported bug happened because of loss of double precision when
converting from double to string and then back to double. As a result,
the  methods to find the index for a particular time value ended up
being off by 1 between the multiple instances of pqAnimationTimeWidget.
This fixes that by using maximum precision of 17 for timesteps. Also
updated the xmls to ensure that state files save the same precision for
timesteps.