Fixes various issues with pqColorAnnotationsPropertyWidget.
This fixes BUG #15483 and BUG #15481. pqColorAnnotationsPropertyWidget
tried to keep the "Annotations" and "IndexedColors" separate from one
another (as it is unfortunately done on the Proxy). That was causing all
the confusion. Hence, ditched that separation. Now Annotations and the
associated colors are tracked together with one another resolving the
reported issues.

Also updated tests since this affects the scalar bars being rendered
after annotation presets are loaded.