BUG #15539: Hack to fix warning message.
The warning message reported by the BUG pops up since the ColorArrayName
property was incorrectly inherited when InheritRepresentationProperties
setting is set to 1. Using a temporary fix to avoid inheriting the value
for scalar coloring related properties. We'll have to devise a more
generic strategy as explained in the inline comments.