BUG #15586: Fixes updating of ColorArrayName.
"ColorArrayName" property on a composite representation wasn't updated
proeprty when representation type changed, except to the arrays
available on the subproxy from which the ColorArrayName property was
exposed.

Since this is a unique requirement for vtkSMRepresentedArrayListDomain,
we add a capability to vtkSMRepresentedArrayListDomain to monitor the
composite-representation proxy for represented data information.