Don't assume active representation/view exist
This fixes a segfault in custom application code when a color editor is
present without an active representation by adding checks to the calls
to renderViewEventually and substituting calls to alternative functions
such as getting the active view and calling render on it.

This also disables buttons that depend on active representation when the
active representation does not exist.