improve material property specification
User can now provide a new material properties array for pairs
of materials (ie interfaces). If they do not the mapper will
try to use parent materials to construct one. This construction
makes solids win over dielectrics, merges the characteristics
of dielectrics to sort out the indices of refraction, and if those
fail arbitrarily picks the properties of the highest material
parent.