Fixed segfault on OsX when swapping empty frame.
The root cause of this issue was that we were deleting pqViewFrame in a
callback while QDrag object was in its start() call. That was causing
the QDrag object to be destroyed since the QDrag object's parent was the
pqViewFrame that was deleted. This fix delays the swapping of frames
until the QDrag operation has completed, thus avoid this segfault.