Fix a crash in the plugin loader.
The current behaviour was to register all dynamic libraries that were
opened as potential plugins. If some were not there were close but
their reference was keeped in the plugin list of libraries to close
at application exit.
The plugin library registration is now moved when we are sure that
the library is a correct plugin.