BUG #15486: Release reference in Accessor.
smtrace creates accessor was each proxy object it traces. We were
missing a reference release in finalize that was causing traced objects
to be held on to for longer than needed.

Since the view was alive, a timer event would trigger a render on it and
that would cause the reported bug with the chart window popping out.