add a convenient way to read material arrays
The new filter passes through its input dataobject but adds to it
the field data arrays of the provided vtk data file. The added
arrays replaces arrays of the same names on the input, but leaves
others untouched.

I chose vtk format only because it was expedient to do so.