Add hints to show/hide toolbar buttons in view frame
View proxies can now add hints to control which toolbar buttons in the
view frame are visible/active.

<Hints>
  <StandardViewFrameActions default_actions="none" />
</Hints>

disables all toolbar buttons while

<Hints>
  <StandardViewFrameActions default_actions="none" />
  <ToggleInteractionMode />
</Hints>

disables all toolbar buttons except for the button that toggles
between 2D/3D camera interactions. This type of hint can be used
to add only the toolbar buttons that are desired for the view.

To show all the default toolbar buttons except for the button that
brings up the camera control dialog, use

<Hints>
  <StandardViewFrameActions>
    <AdjustCamera visibility="never" />
  </StandardViewFrameActions>
</Hints>