Update point gaussian plugin to leverage new mapper features.
This commit exposes new features recently added to the OpenGL2
vtkOpenGLPointGaussianMapper class, such as the ability to set
sprite color, opacity, and scale by different arrays.  Also
exposed is the abilty to map array values to opacity and scale using
vtkpiecewisefunction.  Additionally, a custom fragment shader snippet
can be provided in order to achieve unique visual effects on the
sprites.