BUG: 13130 fixed scalar bar settings
The AutoOrient and Orientation seetetings were not being restored from
the settings file. Since these have an information property, they were
being reset to the value of the information property. Added a
<NoDefault /> hint to stop this from happening.

Also, changed the position of scalar bars oriented horizontally
when they are initially added to keep them in the view.