Fixes BUG #15645.
The reported issue was arising because of the fact the pqSpinBox would
fire editingFinished() event when it is disabled while it had the focus.
Since disabling a widget makes it loose the focus, it would fire the
editingFinished() signal. pqAnimationTimeWidget was observing that
signal and then trying to change the time on a non-existant TimeKeeper
proxy and hence the segfault.

The fix extends pqSpinBox to add new signal
valueChangedAndEditingFinished (similar in spirit to what we do with
pqLineEdit) to limit the overzealous firing of editingFinished() signal.
That addresses the reported segfault.