Fix bug and inefficiency in request data.
The application of the mask filter was not necessary in the case of
polydata input.  And since the filter was specified to create one
cell per point, if a polydata came in with a subset of the points
as cells, that subsetting was destroyed by the filter.  Also, i
resulted in doubling of the memory required to represent cell
points, at least in the case where the input had a single cell with
all the indices as vertices.

Now the mask filter is only applied to unstructured grid input, and
no cells are created in that case, as the underlying mapper handles
the points just fine.