BUG #15638: Updating a reader does not update label array name.
Certain readers allow you to select fields you want to load from the file.
If initially all fields are loaded, we select one of them as array to label with,
and then reload the data without that field and error is produced.

This is fixed by listening to dataUpdated() and removing the label if the array
was removed from the data.