BUG 14388: Fixed bug when text is pasted in with mouse middle-click
Middle-mouse-clicking to paste text into the Python console was
broken. The problem was that text was being inserted into the document
before the return key was pressed, so the position of the cursor was
effectively at the end of the pasted text. Hence, the console thought
there was no command to execute.

Fixed this by overiding mouseReleaseEvent(QMouseEvent* e) and updating
the correct item in the CommandHistory member variable with the pasted
text.